loadFinchConfig function is only called on macOS
